Update echo.php
The <a href="https://github.com/bwssytems/ha-bridge">ha-bridge</a> supports dim commands from the Echo. The bridge can pass variables based on your spoken requested percentage. They are ${intensity.percent} (for 1-100) or ${intensity.byte} (for 0-255). By saying, “Alexa, dim the kitchen lights to 50%” that percentage will be passed in the ‘On URL’.